requirements.txt
ultralytics==8.0.153
torch>=1.7.0
pyyaml